backend/requirements.txt
# Django y REST
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2

# Base de datos
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Utilidades
python-dateutil==2.9.0
python-dotenv==1.0.1

# Archivos e imágenes
Pillow==10.4.0
openpyxl==3.1.5

# Almacenamiento S3
boto3==1.35.36
django-storages==1.14.4

# Producción
gunicorn==22.0.0
whitenoise==6.7.0

# IA - Claude API
httpx==0.27.2
anthropic==0.34.2

# PDF processing
pdfplumber==0.11.4
PyPDF2==3.0.1
reportlab==4.2.2

# Async support
asgiref==3.8.1
